Escape regex search patterns and enable custom include macros
Both search word and context should only contain maintext words so on
comparisons of search and match phrases both should first be only cleaned
maintext words.

The user can supply macros that should be preserved in the maintext
clean function.

This solves issue #7 and issue #8.

This includes refactoring of Macro and Bracket objects for handling latex
macros.

The Macro object is new, making some methods available that will return
the content of the macro or the complete macro, as well as attributes given name
of the macro as well as all its content before opening bracket (if present).
The Brackets object is the old Macro object handling parsing of latex brackets
content and giving its length.

This also replaces the match of `\b` with negative look-arounds in the regex,
cf. https://stackoverflow.com/questions/45145626/python-re-escape-and-r-b-gives-unexpected-results

The argument lets user add file with macros that should be included in text
comparison.