Basic functionality for issue-8 working
This includes some major internal changes.
- Added data structure for string cleanup, Word, which registers cleaned words
and their non-cleaned positions in the original string. This makes matching in
normalized versions possible while still being able to map it onto the original
and messy string.
- Rewritten clean function using that Word object. A much more complex function
than previously. This returns a nested list of word objects. When samewords
annotations are contained in the string being cleaned a nested Word object is
created to maintain clean version and keeping information on the sameword wrap.
- Add flatten_word_list function to make the nested list available to regex
searching.
- Rewritten 'replace_in_string' function for the new Word object.
- Rewritten 'wrap_phrase' function for the new Word object.
- Tests